Enhance validation rules and update volume data type in SpesifikasiKerangkaAcuan

+++ database/migrations/2024_09_15_075202_create_spesifikasi_kerangka_acuans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('spesifikasi_kerangka_acuans', function (Blueprint $table) {
             $table->id();
             $table->string('rincian', 80)->nullable();
-            $table->integer('volume')->nullable()->unsigned();
+            $table->decimal('volume', 10, 2)->nullable()->unsigned();
             $table->string('satuan', 40)->nullable();
             $table->integer('harga_satuan')->nullable()->unsigned();
             $table->integer('total_harga')->nullable()->unsigned();
